% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VBJM}
\alias{VBJM}
\title{Main function to run VBJM}
\usage{
VBJM(datalist, paralist, maxiter = 100L, eps = 1e-04)
}
\arguments{
\item{datalist}{a list object with all data}

\item{paralist}{a list object with all initial parameters}

\item{maxiter}{the maximum number of iterations}

\item{eps}{stopping criteria}
}
\value{
a list
}
\description{
Main function to run VBJM
}
